requirements.txt
python-dotenv
requests
anthropic
Appium-Python-Client
